+++ src/providers/bash.ts
@@ -6,32 +6,13 @@ import * as readdir from 'readdir-enhanced';
 import * as bglob from 'bash-glob';
 import * as micromatch from 'micromatch';
 
-import { ITask } from './task';
 import { IOptions } from '../fglob';
-
-function fileStat(filepath: string): Promise<readdir.IEntry> {
-	return new Promise((resolve, reject) => {
-		fs.stat(filepath, (err, stat) => {
-			if (err) {
-				return reject(err);
-			}
-			resolve(stat);
-		});
-	});
-}
-
-function isNegative(pattern: string): boolean {
-	return pattern[0] === '!';
-}
-
-function getNegativeAsPositivePatterns(patterns: string[]) {
-	return patterns
-		.filter((pattern) => isNegative(pattern))
-		.map((pattern) => pattern.slice(1));
-}
+import { ITask } from '../utils/task';
+import { statFile } from '../utils/io';
+import { isNegative, getNegativeAsPositive } from '../utils/patterns';
 
 function filterByNegativePatterns(files: string[], patterns: string[]) {
-	const negativePatterns = getNegativeAsPositivePatterns(patterns);
+	const negativePatterns = getNegativeAsPositive(patterns);
 	const entries = [];
 	for (let i = 0; i < files.length; i++) {
 		if (micromatch(files[i], negativePatterns).length === 0) {
@@ -53,7 +34,7 @@ export function async(task: ITask, options: IOptions): Promise<string[] | readdi
 
 			const entries = filterByNegativePatterns(files, task.patterns);
 			if (options.stats || options.onlyFiles || options.onlyDirs) {
-				return Promise.all(entries.map(fileStat)).then((stats: any) => {
+				return Promise.all(entries.map(statFile)).then((stats: readdir.IEntry[]) => {
 					const results: (string | readdir.IEntry)[] = [];
 
 					for (let i = 0; i < stats.length; i++) {
@@ -61,14 +42,13 @@ export function async(task: ITask, options: IOptions): Promise<string[] | readdi
 						if ((options.onlyFiles && !entry.isFile()) || (options.onlyDirs && !entry.isDirectory())) {
 							continue;
 						}
-
 						entry.path = entries[i];
 						results.push(cb(options.stats ? entry : entry.path));
 					}
 					resolve(results);
 				});
 			}
-			resolve(entries.map(cb));
+			resolve(options.transform ? (<any>entries).map(cb) : entries);
 		});
 	});
 }
@@ -91,7 +71,7 @@ export function sync(task: ITask, options: IOptions): (string | readdir.IEntry)[
 			entry.path = entries[i];
 			results.push(cb(options.stats ? entry : entry.path));
 		}
-		return results.map(cb);
+		return results;
 	}
-	return entries.map(cb);
+	return options.transform ? entries.map(cb) : entries;
 }
